Refactor(modal): movendo para pasta personalizada de modal
