--- src/bot/modals/compet/createMemberEmbed.json
@@ -1,6 +1,0 @@
-{
-    "color": "#f56565",
-    "title": "Não foi possível criar o novo membro",
-    "description": "Abaixo você encontra as informações sobre o erro",
-    "thumbnail": "https://www.pngfind.com/pngs/m/0-1420_red-cross-mark-clipart-green-checkmark-red-x.png"
-}
