--- src/bot/modals/compet/createMemberInputFields.json
@@ -1,34 +1,0 @@
-{
-    "inputFields": [
-        {
-          "label" : "Qual o nome do novo membro?",
-          "placeholder" : "Sandro Renato Dias",
-          "customId" : "nome",
-          "required" : true,
-          "minLength" : 3,
-          "maxLength" : 161,
-          "style" : 1
-        },
-        {
-          "label" : "Qual o email do novo membro?",
-          "placeholder" : "Andrejogadorcarodacruz@gmail.com",
-          "customId" : "email",
-          "required" : true,
-          "style" : 1
-        },
-        {
-          "label" : "Qual o linkedin do novo membro?",
-          "placeholder" : "https://www.linkedin.com/in/NataliaBatista",
-          "customId" : "linkedin",
-          "required" : false,
-          "style" : 1
-        },
-        {
-          "label" : "O link da foto do novo membro",
-          "placeholder" : "https://i.ibb.co/j4PDk7n/btg3.jpg",
-          "customId" : "url_imagem",
-          "required" : false,
-          "style" : 1
-        }
-    ]
-}
