+++ src/bot/modals/compet/createMemberModal/createMemberErrorEmbed.json
