+++ src/bot/modals/compet/createMemberModal/createMemberSuccessEmbed.json
