--- src/bot/modals/compet/modalBuilderRequest.json
@@ -1,4 +1,0 @@
-{
-  "title": "Adicione o mais novo membro do compet!",
-  "customId": "addmember"
-}
